Reorganize tests for server and browser testing
