+++ packages/storefront-api-client/src/tests/validations/validations.server.test.ts
@@ -1,0 +1,11 @@
+import { validatePrivateAccessTokenUsage } from "../../validations";
+
+describe("validatePrivateAccessTokenUsage(): Server", () => {
+  it("does not throw an error when only the private token is provided when within a server environment", () => {
+    const privateAccessToken = "private-token";
+
+    expect(() =>
+      validatePrivateAccessTokenUsage(privateAccessToken)
+    ).not.toThrow();
+  });
+});
